Version 1 done.
The expected basic functions of a simple calculator work as expected. It is not clean, some events are not being handles, such as replacing the last equation's result with the new equation's first number instead of appending to it.

Since using eval() is too resource intensive for such simple calculations the next step will be making my own easier-on-the-system calculator functions. Along the way some of the currently missing events may also be implemented, however that may just as well remain to be done afterwards instead.